[bf-fix] 임시
- diskProfile 이 누락 된 디스크이미지 유형에 대한 처리
  참고: VM hibernation memory 이거나 snapshot memory를 파생적으로 만들어 질 때, 이 저장장치들은 diskProfile을 지니지 못해 follow 검색이 안됨
- [f-mod] 로그인 화면 정정
- [f-fix] 통지함 높이 유동적 변경기능
- [f-mod] 선택해제 기능 (VM 스냅샷, 네트워크 인터페이스)
- [f-add] 스냅샷 동작기능 추가 (모달 및 API연결)

+++ back/rutil-vm-api/src/main/java/com/itinfo/rutilvm/api/controller/computing/VmController.kt
@@ -1258,6 +1258,7 @@ class VmController: BaseController() {
 		log.info("/computing/vms/{}/snapshots/undo ... 가상머신 커밋", vmId)
 		return ResponseEntity.ok(iVmSnapshot.undoFromVm(vmId))
 	}
+// TODO: 보류 (2차개발)
 /*
 	@ApiOperation(
 	 	httpMethod="POST",
@@ -1375,7 +1376,7 @@ class VmController: BaseController() {
 	@ApiImplicitParams(
 		ApiImplicitParam(name="vmId", value="가상머신 ID", dataTypeClass=String::class, required=true, paramType="path"),
 		ApiImplicitParam(name="graphicConsoleId", value="그래픽 콘솔 ID", dataTypeClass=String::class, required=true, paramType="path"),
-		ApiImplicitParam(name="expiry", value="콘솔 접근 만료시간 (초)", dataTypeClass=Int::class, defaultValue="7200", required=true, paramType="query"),
+		ApiImplicitParam(name="expiry", value="콘솔 접근 만료시간 (초)", dataTypeClass=Long::class, defaultValue="7200", required=true, paramType="query"),
 	)
 	@ApiResponses(
 		ApiResponse(code = 201, message = "CREATED"),
@@ -1387,7 +1388,7 @@ class VmController: BaseController() {
 	fun publishTicket(
 		@PathVariable vmId: String? = null,
 		@PathVariable graphicConsoleId: String? = null,
-		@RequestParam expiry: Int? = 7200,
+		@RequestParam expiry: Long? = 7200L,
 	): ResponseEntity<TicketVo> {
 		if (vmId.isNullOrEmpty())
 			throw ErrorPattern.VM_ID_NOT_FOUND.toException()
